Components: Extract PostExcerpt as a reusable component
